+++ packages/connectivity-controller/src/index.ts
@@ -6,7 +6,7 @@ export type {
   ConnectivityControllerEvents,
   ConnectivityControllerMessenger,
 } from './ConnectivityController';
-export type { ConnectivityControllerSetStatusAction } from './ConnectivityController-method-action-types';
+export type { ConnectivityControllerSetConnectivityStatusAction } from './ConnectivityController-method-action-types';
 export type { ConnectivityAdapter, ConnectivityStatus } from './types';
 export { CONNECTIVITY_STATUSES } from './types';
 export {